+++ src/java/Controller/Auth/ResetPasswordServlet.java
@@ -4,7 +4,7 @@ import Exception.NotFoundException;
 import Model.Entity.OAuth.PasswordResetToken;
 import Model.Entity.User.User;
 import Service.Auth.PasswordResetTokenService;
-import Service.MailService;
+import Service.External.MailService;
 import Service.User.UserService;
 import java.io.IOException;
 import java.io.PrintWriter;
