+++ src/java/Service/External/CloudinaryService.java
@@ -1,4 +1,4 @@
-package Service;
+package Service.External;
 
 import com.cloudinary.Cloudinary;
 import com.cloudinary.utils.ObjectUtils;
